// frontend/src/containers/pages/MainPage/interface.ts
export interface ListItems {
  firstItem: {
    heading?: string;
    description?: string;
  },
  secondItem: {
    heading?: string;
    description?: string;
  },
  thirdItem: {
    heading?: string;
    description?: string;
  },
  fourthItem: {
    heading?: string;
    description?: string;
  }
}